# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Graphify Knowledge Agent API
  description: Authenticated multi-project API with named SSE conversation streams.
  version: 1.0.0-poc
paths:
...
  /api/v1/plsql/callees:
    get:
      tags:
      - plsql
      summary: List Callees
      description: Return routines called by the requested routine (outgoing CALLS).
      operationId: list_callees_api_v1_plsql_callees_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlDependencyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/callers:
    get:
      tags:
      - plsql
      summary: List Callers
      description: Return routines that call the requested routine (incoming CALLS).
      operationId: list_callers_api_v1_plsql_callers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlDependencyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/files:
    get:
      tags:
      - plsql
      summary: Get File Source
      description: |-
        Return read-only source text for one file of the analyzed corpus.

        ``startLine``/``endLine`` are optional request ranges echoed as the
        highlight for the viewer. Unknown file ids and traversal attempts are
        rejected as ``analysis_not_found``; oversized files are capped by
        ``plsql_max_source_bytes``.
      operationId: get_file_source_api_v1_plsql_files_get
      security:
      - HTTPBearer: []
      parameters:
      - name: fileId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Fileid
      - name: startLine
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Startline
      - name: endLine
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Endline
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlSourceContent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/impact:
    get:
      tags:
      - plsql
      summary: Get Impact
      description: |-
        Return bounded transitive dependents of a changed object.

        Dependents reach the changed object over typed dependency relationships
        (``CALLS | READS | WRITES | VIEW_DEPENDS_ON``) within
        ``plsql_max_hops``; items are ordered by distance and then lexicographic
        ids, and every item carries its shortest explaining path(s) with per-hop
        evidence. No severity is computed or persisted.
      operationId: get_impact_api_v1_plsql_impact_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlImpactResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/object:
    get:
      tags:
      - plsql
      summary: Get Object
      description: 'Return one analyzed PL/SQL object by its opaque identifier.


        Identifiers embed ``/`` characters (``plsql://...``), so the identifier

        travels as a query parameter instead of a path segment to survive

        proxies and routers without double encoding.'
      operationId: get_object_api_v1_plsql_object_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlObject'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/objects:
    get:
      tags:
      - plsql
      summary: Search Objects
      description: Deterministically search analyzed PL/SQL objects (bounded).
      operationId: search_objects_api_v1_plsql_objects_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
          default: ''
          title: Q
      - name: kinds
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              enum:
              - Table
              - View
              - Package
              - Sequence
              - Trigger
              - Index
              - Synonym
              - Type
              - Procedure
              - Function
              - AnonymousBlock
              type: string
          - type: 'null'
          title: Kinds
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlObjectSearchResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/paths:
    get:
      tags:
      - plsql
      summary: Find Paths
      description: |-
        Return bounded dependency paths from one object to another.

        Paths traverse typed dependency relationships (``CALLS | READS | WRITES |
        VIEW_DEPENDS_ON``) within ``plsql_max_hops``, ordered by hop count then
        lexicographic node ids, with duplicates collapsed and truncation reported
        when the row cap is hit.
      operationId: find_paths_api_v1_plsql_paths_get
      security:
      - HTTPBearer: []
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: To
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlPathResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/relationships/evidence:
    get:
      tags:
      - plsql
      summary: Get Relationship Evidence
      description: |-
        Return one typed relationship with its source evidence coordinates.

        Identifiers embed ``/`` characters (``edge://...``), so the identifier
        travels as a query parameter instead of a path segment.
      operationId: get_relationship_evidence_api_v1_plsql_relationships_evidence_get
      security:
      - HTTPBearer: []
      parameters:
      - name: relationshipId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Relationshipid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlDependency'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/source:
    get:
      tags:
      - plsql
      summary: Get Object Source
      description: |-
        Return read-only source text for an analyzed object.

        The response highlights the object's declaration line. Content is served
        strictly under the configured ``plsql_source_root`` with traversal guards
        and the configured byte cap.
      operationId: get_object_source_api_v1_plsql_source_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlSourceContent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/table-access:
    get:
      tags:
      - plsql
      summary: List Table Access
      description: Return READS/WRITES/TRIGGER_ON/VIEW_DEPENDS_ON edges of an object.
      operationId: list_table_access_api_v1_plsql_table_access_get
      security:
      - HTTPBearer: []
      parameters:
      - name: objectId
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 512
          title: Objectid
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlDependencyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plsql/unresolved:
    get:
      tags:
      - plsql
      summary: List Unresolved
      description: |-
        Return edges whose resolution is AMBIGUOUS or UNRESOLVED.

        Uncertainty is surfaced as data: these edges are never presented as
        certain relationships by the console.
      operationId: list_unresolved_api_v1_plsql_unresolved_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlsqlDependencyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlsqlDependency:
      properties:
        id:
          type: string
          maxLength: 512
          minLength: 1
          title: Id
        relationship:
          type: string
          enum:
          - CALLS
          - READS
          - WRITES
          - VIEW_DEPENDS_ON
          - TRIGGER_ON
          - INDEXES
          - SYNONYM_FOR
          - DECLARES
          - CONTAINS
          title: Relationship
        source:
          $ref: '#/components/schemas/PlsqlObjectReference'
        target:
          $ref: '#/components/schemas/PlsqlObjectReference'
        resolution:
          type: string
          enum:
          - EXACT
          - INFERRED
          - AMBIGUOUS
          - UNRESOLVED
          title: Resolution
        evidence:
          anyOf:
          - $ref: '#/components/schemas/PlsqlSourceCoordinate'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - id
      - relationship
      - source
      - target
      - resolution
      title: PlsqlDependency
      description: One typed dependency edge with resolution and source evidence.
    PlsqlDependencyResult:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PlsqlDependency'
          type: array
          title: Items
        truncated:
          type: boolean
          title: Truncated
        count:
          type: integer
          minimum: 0.0
          title: Count
      additionalProperties: false
      type: object
      required:
      - items
      - truncated
      - count
      title: PlsqlDependencyResult
...
    PlsqlImpactResult:
      properties:
        object:
          $ref: '#/components/schemas/PlsqlObjectReference'
        items:
          items:
            $ref: '#/components/schemas/PlsqlImpactItem'
          type: array
          title: Items
        truncated:
          type: boolean
          title: Truncated
        count:
          type: integer
          minimum: 0.0
          title: Count
      additionalProperties: false
      type: object
      required:
      - object
      - items
      - truncated
      - count
      title: PlsqlImpactResult
      description: |-
        Deterministic, bounded impact report for one changed object.

        ``object`` identifies the analyzed object; ``items`` are its transitive
        dependents ordered by distance and then lexicographic ids.
    PlsqlObject:
      properties:
        id:
          type: string
          maxLength: 512
          minLength: 1
          title: Id
        kind:
          type: string
          enum:
          - Table
          - View
          - Package
          - Sequence
          - Trigger
          - Index
          - Synonym
          - Type
          - Procedure
          - Function
          - AnonymousBlock
          title: Kind
        name:
          type: string
          maxLength: 256
          minLength: 1
          title: Name
        schema:
          type: string
          maxLength: 128
          minLength: 1
          title: Schema
        qualifiedName:
          type: string
          maxLength: 512
          minLength: 1
          title: Qualifiedname
        projectId:
          type: string
          maxLength: 128
          minLength: 1
          title: Projectid
        owner:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Owner
        signature:
          anyOf:
          - type: string
            maxLength: 1024
          - type: 'null'
          title: Signature
        returnType:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Returntype
        declaration:
          anyOf:
          - $ref: '#/components/schemas/PlsqlSourceCoordinate'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - id
      - kind
      - name
      - schema
      - qualifiedName
      - projectId
      title: PlsqlObject
      description: Object summary/detail returned by search and object endpoints.
...
    PlsqlObjectSearchResult:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PlsqlObject'
          type: array
          title: Items
        truncated:
          type: boolean
          title: Truncated
        count:
          type: integer
          minimum: 0.0
          title: Count
      additionalProperties: false
      type: object
      required:
      - items
      - truncated
      - count
      title: PlsqlObjectSearchResult
...
    PlsqlPathResult:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PlsqlPath'
          type: array
          title: Items
        truncated:
          type: boolean
          title: Truncated
        count:
          type: integer
          minimum: 0.0
          title: Count
      additionalProperties: false
      type: object
      required:
      - items
      - truncated
      - count
      title: PlsqlPathResult
      description: Deterministic, bounded envelope of dependency paths.
    PlsqlSourceContent:
      properties:
        file:
          $ref: '#/components/schemas/PlsqlSourceFile'
        lines:
          items:
            type: string
          type: array
          title: Lines
        highlight:
          anyOf:
          - $ref: '#/components/schemas/PlsqlSourceHighlight'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - file
      - lines
      title: PlsqlSourceContent
      description: |-
        Read-only file content with an optional highlight range.

        ``lines`` holds the whole file split at line breaks; the viewer renders
        line numbers itself. Raw text is served only through this endpoint,
        scoped strictly under the configured source root.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer